Flatten package structure, removed Person, NamedEntity and BaseEntity super classes
