--- spring-petclinic-customers-service/src/main/java/org/springframework/samples/petclinic/customers/domain/model/pet/Pet.java
@@ -1,73 +1,0 @@
-/*
- * Copyright 2002-2013 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.springframework.samples.petclinic.customers.domain.model.pet;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import org.springframework.samples.petclinic.customers.domain.model.owner.Owner;
-import org.springframework.samples.petclinic.customers.support.jpa.NamedEntity;
-
-import javax.persistence.*;
-import java.util.Date;
-
-/**
- * Simple business object representing a pet.
- *
- * @author Ken Krebs
- * @author Juergen Hoeller
- * @author Sam Brannen
- * @author Maciej Szarlinski
- */
-@Entity
-@Table(name = "pets")
-public class Pet extends NamedEntity {
-
-    @Column(name = "birth_date")
-    @Temporal(TemporalType.DATE)
-    private Date birthDate;
-
-    @ManyToOne
-    @JoinColumn(name = "type_id")
-    private PetType type;
-
-    @ManyToOne
-    @JoinColumn(name = "owner_id")
-    @JsonIgnore
-    private Owner owner;
-
-    public void setBirthDate(Date birthDate) {
-        this.birthDate = birthDate;
-    }
-
-    public Date getBirthDate() {
-        return birthDate;
-    }
-
-    public void setType(PetType type) {
-        this.type = type;
-    }
-
-    public PetType getType() {
-        return type;
-    }
-
-    public void setOwner(Owner owner) {
-        this.owner = owner;
-    }
-
-    public Owner getOwner() {
-        return owner;
-    }
-}
