+++ spring-petclinic-customers-service/src/main/java/org/springframework/samples/petclinic/customers/web/OwnerResource.java
@@ -1,0 +1,90 @@
+/*
+ * Copyright 2002-2013 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.samples.petclinic.customers.web;
+
+import lombok.RequiredArgsConstructor;
+
+import java.util.List;
+
+import javax.validation.Valid;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.samples.petclinic.customers.model.Owner;
+import org.springframework.samples.petclinic.customers.model.OwnerRepository;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * @author Juergen Hoeller
+ * @author Ken Krebs
+ * @author Arjen Poutsma
+ * @author Michael Isvy
+ * @author Maciej Szarlinski
+ */
+@RequestMapping("/owners")
+@RestController
+@RequiredArgsConstructor(onConstructor = @__(@Autowired))
+class OwnerResource {
+
+    private final OwnerRepository ownerRepository;
+
+    /**
+     * Create Owner
+     */
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public void createOwner(@Valid @RequestBody Owner owner) {
+        ownerRepository.save(owner);
+    }
+
+    /**
+     * Read single Owner
+     */
+    @GetMapping(value = "/{ownerId}")
+    public Owner findOwner(@PathVariable("ownerId") int ownerId) {
+        return ownerRepository.findOne(ownerId);
+    }
+
+    /**
+     * Read List of Owners
+     */
+    @GetMapping
+    public List<Owner> findAll() {
+        return ownerRepository.findAll();
+    }
+
+    /**
+     * Update Owner
+     */
+    @PutMapping(value = "/{ownerId}")
+    public Owner updateOwner(@PathVariable("ownerId") int ownerId, @Valid @RequestBody Owner ownerRequest) {
+        final Owner ownerModel = ownerRepository.findOne(ownerId);
+        // This is done by hand for simplicity purpose. In a real life use-case we should consider using MapStruct.
+        ownerModel.setFirstName(ownerRequest.getFirstName());
+        ownerModel.setLastName(ownerRequest.getLastName());
+        ownerModel.setCity(ownerRequest.getCity());
+        ownerModel.setAddress(ownerRequest.getAddress());
+        ownerModel.setTelephone(ownerRequest.getTelephone());
+        return ownerRepository.save(ownerModel);
+    }
+}
