+++ spring-petclinic-customers-service/src/main/java/org/springframework/samples/petclinic/customers/web/PetResource.java
@@ -1,0 +1,91 @@
+/*
+ * Copyright 2002-2013 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.samples.petclinic.customers.web;
+
+import lombok.RequiredArgsConstructor;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.samples.petclinic.customers.model.Owner;
+import org.springframework.samples.petclinic.customers.model.OwnerRepository;
+import org.springframework.samples.petclinic.customers.model.Pet;
+import org.springframework.samples.petclinic.customers.model.PetRepository;
+import org.springframework.samples.petclinic.customers.model.PetType;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * @author Juergen Hoeller
+ * @author Ken Krebs
+ * @author Arjen Poutsma
+ * @author Maciej Szarlinski
+ */
+@RestController
+@RequiredArgsConstructor(onConstructor = @__(@Autowired))
+class PetResource {
+
+    private final PetRepository petRepository;
+
+    private final OwnerRepository ownerRepository;
+
+    @GetMapping("/petTypes")
+    public List<PetType> getPetTypes() {
+        return petRepository.findPetTypes();
+    }
+
+    @PostMapping("/owners/{ownerId}/pets")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void processCreationForm(
+        @RequestBody PetRequest petRequest,
+        @PathVariable("ownerId") int ownerId) {
+
+        final Pet pet = new Pet();
+        final Owner owner = ownerRepository.findOne(ownerId);
+        owner.addPet(pet);
+
+        save(pet, petRequest);
+    }
+
+    @PutMapping("/owners/*/pets/{petId}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void processUpdateForm(@RequestBody PetRequest petRequest) {
+        save(petRepository.findOne(petRequest.getId()), petRequest);
+    }
+
+    private void save(final Pet pet, final PetRequest petRequest) {
+
+        pet.setName(petRequest.getName());
+        pet.setBirthDate(petRequest.getBirthDate());
+
+        petRepository.findPetTypeById(petRequest.getTypeId())
+            .ifPresent(pet::setType);
+
+        petRepository.save(pet);
+    }
+
+    @GetMapping("owners/*/pets/{petId}")
+    public PetDetails findPet(@PathVariable("petId") int petId) {
+        return new PetDetails(petRepository.findOne(petId));
+    }
+
+}
