+++ spring-petclinic-vets-service/src/main/java/org/springframework/samples/petclinic/vets/system/VetsProperties.java
@@ -1,0 +1,25 @@
+package org.springframework.samples.petclinic.vets.system;
+
+import lombok.Data;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+/**
+ * Typesafe custom configuration.
+ *
+ * @author Maciej Szarlinski
+ */
+@Data
+@ConfigurationProperties(prefix = "vets")
+public class VetsProperties {
+
+    private Cache cache;
+
+    @Data
+    public static class Cache {
+
+        private int ttl;
+
+        private int heapSize;
+    }
+}
