--- spring-petclinic-vets-service/src/main/java/org/springframework/samples/petclinic/vets/web/boundary/VetResource.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 2002-2013 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.springframework.samples.petclinic.vets.web.boundary;
-
-import lombok.RequiredArgsConstructor;
-
-import java.util.List;
-
-import org.springframework.samples.petclinic.vets.domain.model.vet.Vet;
-import org.springframework.samples.petclinic.vets.domain.model.vet.VetRepository;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-/**
- * @author Juergen Hoeller
- * @author Mark Fisher
- * @author Ken Krebs
- * @author Arjen Poutsma
- * @author Maciej Szarlinski
- */
-@RequestMapping("/vets")
-@RestController
-@RequiredArgsConstructor
-class VetResource {
-
-    private final VetRepository vetRepository;
-
-    @GetMapping
-    public List<Vet> showResourcesVetList() {
-        return vetRepository.findAll();
-    }
-}
