--- spring-petclinic-vets-service/src/test/java/org/springframework/samples/petclinic/vets/web/boundary/VetResourceTest.java
@@ -1,47 +1,0 @@
-package org.springframework.samples.petclinic.vets.web.boundary;
-
-import static java.util.Arrays.asList;
-import static org.mockito.BDDMockito.given;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
-import org.springframework.samples.petclinic.vets.domain.model.vet.Vet;
-import org.springframework.samples.petclinic.vets.domain.model.vet.VetRepository;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.servlet.MockMvc;
-
-/**
- * @author Maciej Szarlinski
- */
-@RunWith(SpringRunner.class)
-@WebMvcTest(VetResource.class)
-@ActiveProfiles("test")
-public class VetResourceTest {
-
-    @Autowired
-    private MockMvc mvc;
-
-    @MockBean
-    private VetRepository vetRepository;
-
-    @Test
-    public void shouldGetAListOfVetsInJSonFormat() throws Exception {
-
-        Vet vet = new Vet();
-        vet.setId(1);
-
-        given(vetRepository.findAll()).willReturn(asList(vet));
-
-        mvc.perform(get("/vets.json").accept(MediaType.APPLICATION_JSON))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("$[0].id").value(1));
-    }
-}
