--- spring-petclinic-visits-service/src/main/java/org/springframework/samples/petclinic/visits/domain/model/visit/Visit.java
@@ -1,100 +1,0 @@
-/*
- * Copyright 2002-2013 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.springframework.samples.petclinic.visits.domain.model.visit;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import org.springframework.samples.petclinic.visits.support.jpa.BaseEntity;
-
-import javax.persistence.*;
-import javax.validation.constraints.Size;
-import java.util.Date;
-
-/**
- * Simple JavaBean domain object representing a visit.
- *
- * @author Ken Krebs
- * @author Maciej Szarlinski
- */
-@Entity
-@Table(name = "visits")
-public class Visit extends BaseEntity {
-
-    /**
-     * Holds value of property date.
-     */
-    @Column(name = "visit_date")
-    @Temporal(TemporalType.TIMESTAMP)
-    @JsonFormat(pattern = "yyyy-MM-dd")
-    private Date date = new Date();
-
-    /**
-     * Holds value of property description.
-     */
-    @Size(max = 8192)
-    @Column(name = "description")
-    private String description;
-
-    /**
-     * Holds id of property pet.
-     */
-    @Column(name = "pet_id")
-    private int petId;
-
-    /**
-     * Getter for property date.
-     *
-     * @return Value of property date.
-     */
-    public Date getDate() {
-        return date;
-    }
-
-    /**
-     * Setter for property date.
-     *
-     * @param date New value of property date.
-     */
-    public void setDate(Date date) {
-        this.date = date;
-    }
-
-    /**
-     * Getter for property description.
-     *
-     * @return Value of property description.
-     */
-    public String getDescription() {
-        return description;
-    }
-
-    /**
-     * Setter for property description.
-     *
-     * @param description New value of property description.
-     */
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public int getPetId() {
-        return petId;
-    }
-
-    public void setPetId(final int petId) {
-        this.petId = petId;
-    }
-
-}
